requirements.txt
autopep8==2.0.1
bcrypt==4.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
Flask==2.2.3
Flask-Cors==3.0.10
google-auth==2.16.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
PyYAML==6.0
requests==2.28.2
rsa==4.9
six==1.16.0
tomli==2.0.1
urllib3==1.26.14
Werkzeug==2.2.3
